+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 ### Other changes
 
 - Add 0x prefix to JSON serialized bytes
+- Rename `repl.loadAbi` to `abi.load` and `repl.fetchAbi` to `abi.fetch`
 
 ## v0.1.4 (2025-01-26)